6,8,9 DONE with buttons
