+++ xrayimageproject/xrayimageproject/Form1.cs
@@ -712,7 +712,7 @@ namespace xrayimageproject
         // generating report
         private void guna2Button12_Click_1(object sender, EventArgs e)
         {
-            ReportInputForm inputForm = new();
+            InputForm inputForm = new();
             DialogResult result = inputForm.ShowDialog(); // Show as a modal dialog
 
             if (result == DialogResult.OK)
@@ -723,8 +723,7 @@ namespace xrayimageproject
                 string diagnosisText = inputForm.txtDiagnosis.Text;
                 Image image = pictureBox1.Image;
 
-                ReportGeneration repoGene = new();
-                repoGene.GenerateReport(patientName, id.ToString(), diagnosisText, image);
+                ReportGeneration.GenerateReport(patientName, id.ToString(), diagnosisText, image);
             }
         }
         // compress files and directories
@@ -852,87 +851,7 @@ namespace xrayimageproject
 
             }
         }
-        partial class ReportInputForm : Form
-        {
-            public ReportInputForm() => InitializeComponent();
 
-            private void InitializeComponent()
-            {
-                this.label1 = new System.Windows.Forms.Label();
-                this.txtPatientName = new System.Windows.Forms.TextBox();
-                this.label2 = new System.Windows.Forms.Label();
-                this.txtDiagnosis = new System.Windows.Forms.TextBox();
-                this.btnGenerateReport = new System.Windows.Forms.Button();
-                this.SuspendLayout();
-
-                // label1
-                this.label1.AutoSize = true;
-                this.label1.Location = new System.Drawing.Point(30, 30);
-                this.label1.Name = "label1";
-                this.label1.Size = new System.Drawing.Size(84, 15);
-                this.label1.TabIndex = 0;
-                this.label1.Text = "Patient Name:";
-
-                // txtPatientName
-                this.txtPatientName.Location = new System.Drawing.Point(120, 27);
-                this.txtPatientName.Name = "txtPatientName";
-                this.txtPatientName.Size = new System.Drawing.Size(200, 23);
-                this.txtPatientName.TabIndex = 1;
-
-                // label2
-                this.label2.AutoSize = true;
-                this.label2.Location = new System.Drawing.Point(30, 70);
-                this.label2.Name = "label2";
-                this.label2.Size = new System.Drawing.Size(60, 15);
-                this.label2.TabIndex = 2;
-                this.label2.Text = "Diagnosis:";
-
-                // txtDiagnosis
-                this.txtDiagnosis.Location = new System.Drawing.Point(120, 67);
-                this.txtDiagnosis.Multiline = true;
-                this.txtDiagnosis.Name = "txtDiagnosis";
-                this.txtDiagnosis.Size = new System.Drawing.Size(200, 100);
-                this.txtDiagnosis.TabIndex = 3;
-
-                // btnGenerateReport
-                this.btnGenerateReport.Location = new System.Drawing.Point(120, 180);
-                this.btnGenerateReport.Name = "btnGenerateReport";
-                this.btnGenerateReport.Size = new System.Drawing.Size(120, 30);
-                this.btnGenerateReport.TabIndex = 4;
-                this.btnGenerateReport.Text = "Generate Report";
-                this.btnGenerateReport.UseVisualStyleBackColor = true;
-                this.btnGenerateReport.Click += new System.EventHandler(btnGenerateReport_Click);
-
-                // InputForm
-                this.AutoScaleDimensions = new System.Drawing.SizeF(7F, 15F);
-                this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-                this.ClientSize = new System.Drawing.Size(354, 231);
-                this.Controls.Add(this.btnGenerateReport);
-                this.Controls.Add(this.txtDiagnosis);
-                this.Controls.Add(this.label2);
-                this.Controls.Add(this.txtPatientName);
-                this.Controls.Add(this.label1);
-                this.Name = "InputForm";
-                this.Text = "Enter Patient Information";
-                this.ResumeLayout(false);
-                this.PerformLayout();
-
-            }
-
-            public System.Windows.Forms.Label label1;
-            public System.Windows.Forms.TextBox txtPatientName;
-            public System.Windows.Forms.Label label2;
-            public System.Windows.Forms.TextBox txtDiagnosis;
-            public System.Windows.Forms.Label label3;
-            public System.Windows.Forms.TextBox imgPath;
-            public System.Windows.Forms.Button btnGenerateReport;
-
-            private void btnGenerateReport_Click(object sender, EventArgs e)
-            {
-                this.DialogResult = DialogResult.OK;
-                this.Close();
-            }
-        }
         private void guna2Button7_Click_1(object sender, EventArgs e)
         {
             if (pictureBox1.Image != null)
@@ -997,6 +916,30 @@ namespace xrayimageproject
             {
                 MessageBox.Show("You Have to Import First Image before to Compare with it");
                 return;
+            }
+        }
+
+        // Add Caption on the Image
+        private void guna2Button18_Click(object sender, EventArgs e)
+        {
+            CaptionInputForm captionInputForm = new();
+            DialogResult result = captionInputForm.ShowDialog(); // Show as a modal dialog
+
+            if (result == DialogResult.OK)
+            {
+
+                SaveFileDialog saveFileDialog = new SaveFileDialog();
+                saveFileDialog.Filter = "Image Files(*.jpg; *.jpeg; *.gif; *.bmp; *.png)|*.jpg; *.jpeg; *.gif; *.bmp; *.png";
+                saveFileDialog.RestoreDirectory = true;
+                saveFileDialog.Title = "Save Captioned Image";
+
+                if (saveFileDialog.ShowDialog() == DialogResult.OK)
+                {
+                    string outputPath = saveFileDialog.FileName;
+
+                    // Add caption on the image
+                    CaptionOnImage.SetCaption(pictureBox1.Image, outputPath, captionInputForm.caption.Text);
+                }
             }
         }
     }
